+++ client/src/store/modules/projects.js
@@ -20,18 +20,18 @@ const getters = {
       const params = {
         projectName: project.name,
       }
-      const isStopped = !!project.stoppedAt
+      const isPaused = !!project.pausedAt
       const isFinished = !!project.finishedAt
-      const isStarted = !!project.startedAt && !isStopped
+      const isStarted = !!project.startedAt && !isPaused
       return {
         ...project,
         isStarted,
-        isStopped,
+        isPaused,
         isFinished,
         mdDescription: marked(project.description, { sanitize: true, breaks: true, smartypants: true }),
         startedAtLabel: isStarted ? formatDate(project.startedAt) : '',
         dueAtLabel: isStarted ? formatDate(project.dueAt) : '',
-        stoppedAtLabel: isStopped ? formatDate(project.stoppedAt) : '',
+        pausedAtLabel: isPaused ? formatDate(project.pausedAt) : '',
         finishedAtLabel: isFinished ? formatDate(project.finishedAt) : '',
         finishedTasksCount,
         tasksCount: tasks.length,
@@ -128,9 +128,9 @@ const actions = {
       .then((res) => commit('tasks/startTasksForProject', project.id, { root: true }))
   },
 
-  stop ({ commit }, { project }) {
+  pause ({ commit }, { project }) {
     return projectsApi
-      .stop(project)
+      .pause(project)
       .then((res) => commit('set', res.data))
       .then((res) => commit('tasks/cancelTasksForProject', project.id, { root: true }))
   },