+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 ### Breaking changes
 
 - API now exposes `task.startedAt` instead of `task.createdAt`
+- API now exposes `task.pausedAt` instead of `task.stoppedAt`
 
 ## Apus 2017-11-26